Add NativeOutputWriter
