+++ backends-clickhouse/src/main/java/org/apache/spark/sql/execution/datasources/CHDatasourceJniWrapper.java
@@ -18,11 +18,15 @@ package org.apache.spark.sql.execution.datasources;
 
 public class CHDatasourceJniWrapper {
 
-  public native long nativeInitFileWriterWrapper(
-      String filePath, byte[] preferredSchema, String formatHint);
+  public native void write(long instanceId, long blockAddress);
 
-  public native long nativeInitMergeTreeWriterWrapper(
-      byte[] plan,
+  public native String close(long instanceId);
+
+  /// FileWriter
+  public native long createFilerWriter(String filePath, byte[] preferredSchema, String formatHint);
+
+  /// MergeTreeWriter
+  public native long createMergeTreeWriter(
       byte[] splitInfo,
       String uuid,
       String taskId,
@@ -31,43 +35,28 @@ public class CHDatasourceJniWrapper {
       byte[] confArray);
 
   public native String nativeMergeMTParts(
-      byte[] plan,
-      byte[] splitInfo,
-      String uuid,
-      String taskId,
-      String partition_dir,
-      String bucket_dir);
+      byte[] splitInfo, String uuid, String taskId, String partition_dir, String bucket_dir);
 
   public static native String filterRangesOnDriver(byte[] plan, byte[] read);
 
-  public native void write(long instanceId, long blockAddress);
-
-  public native void writeToMergeTree(long instanceId, long blockAddress);
-
-  public native void close(long instanceId);
-
-  public native String closeMergeTreeWriter(long instanceId);
-
-  /*-
+  /**
    * The input block is already sorted by partition columns + bucket expressions. (check
-   * org.apache.spark.sql.execution.datasources.FileFormatWriter#write)
-   * However, the input block may contain parts(we call it stripe here) belonging to
-   * different partition/buckets.
+   * org.apache.spark.sql.execution.datasources.FileFormatWriter#write) However, the input block may
+   * contain parts(we call it stripe here) belonging to different partition/buckets.
    *
-   * If bucketing is enabled, the input block's last column is guaranteed to be _bucket_value_.
+   * <p>If bucketing is enabled, the input block's last column is guaranteed to be _bucket_value_.
    *
-   * This function splits the input block in to several blocks, each of which belonging
-   * to the same partition/bucket. Notice the stripe will NOT contain partition columns
+   * <p>This function splits the input block in to several blocks, each of which belonging to the
+   * same partition/bucket. Notice the stripe will NOT contain partition columns
    *
-   * Since all rows in a stripe share the same partition/bucket,
-   * we only need to check the heading row.
-   * So, for each stripe, the native code also returns each stripe's first row's index.
-   * Caller can use these indice to get UnsafeRows from the input block,
-   * to help FileFormatDataWriter to aware partition/bucket changes.
+   * <p>Since all rows in a stripe share the same partition/bucket, we only need to check the
+   * heading row. So, for each stripe, the native code also returns each stripe's first row's index.
+   * Caller can use these indices to get UnsafeRows from the input block, to help
+   * FileFormatDataWriter to aware partition/bucket changes.
    */
   public static native BlockStripes splitBlockByPartitionAndBucket(
       long blockAddress,
-      int[] partitionColIndice,
+      int[] partitionColIndices,
       boolean hasBucket,
       boolean reserve_partition_columns);
 }
