+++ cpp-ch/local-engine/tests/gtest_parquet_columnindex.cpp
@@ -20,6 +20,7 @@
 #include <ranges>
 #include <string>
 #include <Columns/ColumnString.h>
+#include <DataTypes/DataTypeString.h>
 #include <IO/ReadBufferFromFile.h>
 #include <Interpreters/ActionsVisitor.h>
 #include <Interpreters/Context.h>
@@ -39,21 +40,21 @@
 #include <Common/BlockTypeUtils.h>
 #include <Common/QueryContext.h>
 
-#    define ASSERT_DURATION_LE(secs, stmt) \
-        { \
-            std::promise<bool> completed; \
-            auto stmt_future = completed.get_future(); \
-            std::thread( \
-                [&](std::promise<bool> & completed) \
-                { \
-                    stmt; \
-                    completed.set_value(true); \
-                }, \
-                std::ref(completed)) \
-                .detach(); \
-            if (stmt_future.wait_for(std::chrono::seconds(secs)) == std::future_status::timeout) \
-                GTEST_FATAL_FAILURE_("       timed out (> " #secs " seconds). Check code for infinite loops"); \
-        }
+#define ASSERT_DURATION_LE(secs, stmt) \
+    { \
+        std::promise<bool> completed; \
+        auto stmt_future = completed.get_future(); \
+        std::thread( \
+            [&](std::promise<bool> & completed) \
+            { \
+                stmt; \
+                completed.set_value(true); \
+            }, \
+            std::ref(completed)) \
+            .detach(); \
+        if (stmt_future.wait_for(std::chrono::seconds(secs)) == std::future_status::timeout) \
+            GTEST_FATAL_FAILURE_("       timed out (> " #secs " seconds). Check code for infinite loops"); \
+    }
 
 
 namespace DB::ErrorCodes
